Added KML warning; colors as parameters in viewer.js
If you try to add a KML or KMZ files to the map, a prompt appears to indicate that they're not yet supported and users are optionally redirected to a service to convert KML/KMZs to a zipped shapefile.  Polygon/polyline styles and colors can be set via parameters in viewer.js.